fix: CI correctness hardening — doctests and mypy now actually run clean
CI was checking less than it appeared to. This patch fixes that without
adding any features:

- pytest was never running the 196 doctests that were the primary
  correctness mechanism throughout development. Added --doctest-modules
  and tests/ to the root pytest config; verified the exact CI command
  now runs all 994 tests+doctests.
- mypy (part of lint.yml) had never been run end-to-end and was failing
  with 59 errors across 23 files. Fixed all of them:
  - Two real bugs: aircraftsim's _aero_and_forces had a wrong return
    type annotation (tuple vs single array); rockettraj's RK4 stepper
    used a generic zip/tuple-comprehension pattern mypy couldn't verify
    preserves 3-tuple arity, rewritten with explicit unpacking and
    numerically verified identical via the existing test suite
  - Added PEP 561 py.typed markers to all 47 packages (none existed)
  - ~28 sites of a confirmed, reproducible typeshed quirk (float**float
    with a non-literal exponent can infer as Any), fixed with explicit
    float()/np.asarray() wraps across 15 packages
  - A missing type annotation on aerounits' shared Pint ureg object
  - 24 remaining errors, all in test files: missing tmp_path: Path
    annotations, missing Optional[float] narrowing before isclose(),
    and type-narrowing at call sites consuming aeroopt's Union-typed
    OptimizationResult.x
- One stray ruff import-ordering issue, auto-fixed

Every fix verified individually (mypy + pytest + doctests) before
moving to the next; final state re-verified as a whole. 994 tests
passing, ruff clean, mypy clean across all 228 source files.

+++ packages/boundarylayer/src/boundarylayer/thickness.py
@@ -48,7 +48,7 @@ def laminar_thickness(x: float, reynolds_x: float) -> float:
 
     """
     _check_reynolds(reynolds_x)
-    return 5.0 * x / reynolds_x**0.5
+    return float(5.0 * x / reynolds_x**0.5)
 
 
 def turbulent_thickness(x: float, reynolds_x: float) -> float:
@@ -68,7 +68,7 @@ def turbulent_thickness(x: float, reynolds_x: float) -> float:
 
     """
     _check_reynolds(reynolds_x)
-    return 0.37 * x / reynolds_x**0.2
+    return float(0.37 * x / reynolds_x**0.2)
 
 
 def laminar_local_skin_friction(reynolds_x: float) -> float:
@@ -81,7 +81,7 @@ def laminar_local_skin_friction(reynolds_x: float) -> float:
 
     """
     _check_reynolds(reynolds_x)
-    return 0.664 / reynolds_x**0.5
+    return float(0.664 / reynolds_x**0.5)
 
 
 def turbulent_local_skin_friction(reynolds_x: float) -> float:
@@ -94,4 +94,4 @@ def turbulent_local_skin_friction(reynolds_x: float) -> float:
 
     """
     _check_reynolds(reynolds_x)
-    return 0.0592 / reynolds_x**0.2
+    return float(0.0592 / reynolds_x**0.2)
